+++ test/units/cli/test_console.py
@@ -20,6 +20,7 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 from ansible.compat.tests import unittest
+from ansible.compat.tests.mock import patch
 
 from ansible.cli.console import ConsoleCLI
 
@@ -30,3 +31,21 @@ class TestConsoleCLI(unittest.TestCase):
         cli.parse()
         self.assertTrue(cli.parser is not None)
 
+    def test_module_args(self):
+        cli = ConsoleCLI([])
+        cli.parse()
+        res = cli.module_args('copy')
+        self.assertTrue(cli.parser is not None)
+        self.assertIn('src', res)
+        self.assertIn('backup', res)
+        self.assertIsInstance(res, list)
+
+    @patch('ansible.utils.display.Display.display')
+    def test_helpdefault(self, mock_display):
+        cli = ConsoleCLI([])
+        cli.parse()
+        cli.modules = set(['copy'])
+        cli.helpdefault('copy')
+        self.assertTrue(cli.parser is not None)
+        self.assertTrue(len(mock_display.call_args_list) > 0,
+                        "display.display should have been called but was not")